requirements.txt
xrpl-py>=2.6.0

